requirements.txt
flask==3.1.2
python-pptx==1.0.2
gunicorn==22.0.0